requirements.txt
python-dotenv
psycopg2-binary
langchain
langchain-gigachat